requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.5.4
python-dotenv==1.0.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9